Simplify: extract comments/reactions group from api.js god-file (slice 1)
First CI-gated slice of the api.js (#7) split. Move the 7 comments/reactions
methods (addComment, listComments, hideComment, removeComment, addReaction,
removeReaction, getReactions) into src/api/comments.js as
createCommentsApi({ refreshSearchIndex }), spread into the api object literal.

Chosen first because it's the cleanest group — the methods touch only
this._getCommentsAutobase (which stays defined in api.js, so `this` still
resolves after the spread) and the single closure helper refreshSearchIndex —
and because backend-test's api-comments-hyperdb.test.mjs actually exercises it.

All call sites invoke these as api.method?.(...) (no destructuring), so `this`
binding is preserved. Verified: node --check on both files, zero leaked closure
identifiers in the new module, keeper methods (_getCommentsAutobase,
getCommentsDebugInfo) retained inline. api.js 4888 -> 4742 lines.
